+++ src/main/java/forestry/api/arboriculture/IArboristTracker.java
@@ -8,5 +8,4 @@ package forestry.api.arboriculture;
 import forestry.api.genetics.IBreedingTracker;
 
 public interface IArboristTracker extends IBreedingTracker {
-
 }