+++ src/main/java/forestry/api/lepidopterology/ILepidopteristTracker.java
@@ -9,7 +9,5 @@ import forestry.api.genetics.IBreedingTracker;
 import forestry.api.lepidopterology.genetics.IButterfly;
 
 public interface ILepidopteristTracker extends IBreedingTracker {
-
 	void registerCatch(IButterfly butterfly);
-
 }