+++ src/main/java/forestry/apiculture/genetics/ApiaristTracker.java
@@ -11,14 +11,11 @@
 package forestry.apiculture.genetics;
 
 import net.minecraft.nbt.CompoundTag;
-import net.minecraft.resources.ResourceLocation;
-import net.minecraft.world.entity.player.Player;
 
 import forestry.api.IForestryApi;
 import forestry.api.apiculture.IApiaristTracker;
 import forestry.api.apiculture.genetics.IBee;
 import forestry.api.genetics.ForestrySpeciesTypes;
-import forestry.api.genetics.IBreedingTracker;
 import forestry.api.genetics.IMutationManager;
 import forestry.api.genetics.ISpecies;
 import forestry.core.genetics.BreedingTracker;
@@ -29,35 +26,31 @@ public class ApiaristTracker extends BreedingTracker implements IApiaristTracker
 	private int dronesTotal = 0;
 	private int princessesTotal = 0;
 
-	/**
-	 * Required for creation from map storage
-	 */
 	public ApiaristTracker() {
-		super();
-	}
-
-	// todo get rid of this constructor and just call load
-	public ApiaristTracker(CompoundTag tag) {
-		super(tag);
+		super(ForestrySpeciesTypes.BEE);
 	}
 
 	@Override
-	protected void load(CompoundTag nbt) {
-		super.load(nbt);
+	public void writeToNbt(CompoundTag nbt) {
+		super.writeToNbt(nbt);
+
+		writeUpdateData(nbt);
+	}
+
+	@Override
+	protected void writeUpdateData(CompoundTag nbt) {
+		nbt.putInt("QueensTotal", this.queensTotal);
+		nbt.putInt("PrincessesTotal", this.princessesTotal);
+		nbt.putInt("DronesTotal", this.dronesTotal);
+	}
+
+	@Override
+	public void readFromNbt(CompoundTag nbt) {
+		super.readFromNbt(nbt);
 
 		this.queensTotal = nbt.getInt("QueensTotal");
 		this.princessesTotal = nbt.getInt("PrincessesTotal");
 		this.dronesTotal = nbt.getInt("DronesTotal");
-	}
-
-	@Override
-	public CompoundTag save(CompoundTag nbt) {
-		nbt.putInt("QueensTotal", this.queensTotal);
-		nbt.putInt("PrincessesTotal", this.princessesTotal);
-		nbt.putInt("DronesTotal", this.dronesTotal);
-
-		nbt = super.save(nbt);
-		return nbt;
 	}
 
 	@Override
@@ -72,7 +65,7 @@ public class ApiaristTracker extends BreedingTracker implements IApiaristTracker
 	@Override
 	public void registerQueen(IBee bee) {
 		this.queensTotal++;
-		// no need to register because we should already have princess
+		registerBirth(bee.getSpecies());
 	}
 
 	@Override
@@ -100,15 +93,5 @@ public class ApiaristTracker extends BreedingTracker implements IApiaristTracker
 	@Override
 	public int getDroneCount() {
 		return this.dronesTotal;
-	}
-
-	@Override
-	protected IBreedingTracker getBreedingTracker(Player player) {
-		return SpeciesUtil.BEE_TYPE.get().getBreedingTracker(player.level, player.getGameProfile());
-	}
-
-	@Override
-	protected ResourceLocation getTypeId() {
-		return ForestrySpeciesTypes.BEE;
 	}
 }